Crate rome_js_parser

Extremely fast, lossless, and error tolerant JavaScript Parser.

The parser uses an abstraction over non-whitespace tokens. This allows us to losslessly or lossly parse code without requiring explicit handling of whitespace. The parser yields events, not an AST, the events are resolved into untyped syntax nodes, which can then be casted into a typed AST.

The parser is able to produce a valid AST from any source code. Erroneous productions are wrapped into ERROR syntax nodes, the original source code is completely represented in the final syntax nodes.

You probably do not want to use the parser struct, unless you want to parse fragments of Js source code or make your own productions. Instead use functions such as parse_script, parse_module, and [] which offer abstracted versions for parsing.

Notable features of the parser are:

Extremely fast parsing and lexing through the extremely fast lexer.

Ability to do Lossy or Lossless parsing on demand without explicit whitespace handling.

Customizable, able to parse any fragments of JS code at your discretion.

Completely error tolerant, able to produce an AST from any source code.

Zero cost for converting untyped nodes to a typed AST.

Ability to go from AST to SyntaxNodes to SyntaxTokens to source code and back very easily with nearly zero cost.

Very easy tree traversal through SyntaxNode.

Descriptive errors with multiple labels and notes.

Very cheap cloning, cloning an ast node or syntax node is the cost of adding a reference to an Rc.

Cheap incremental reparsing of changed text.

The crate further includes utilities such as:

ANSI syntax highlighting of nodes or text through lexer.

It is inspired by the rust analyzer parser but adapted for JavaScript.

Syntax Nodes vs AST Nodes

The crate relies on a concept of untyped rome_js_syntax::JsSyntaxNodes vs typed rome_rowan::AstNodes. Syntax nodes represent the syntax tree in an untyped way. They represent a location in an immutable tree with two pointers. The syntax tree is composed of rome_js_syntax::JsSyntaxNodes and rome_js_syntax::JsSyntaxTokens in a nested tree structure. Each node can have parents, siblings, children, descendants, etc.

rome_rowan::AstNodes represent a typed version of a syntax node. They have the same exact representation as syntax nodes therefore a conversion between either has zero runtime cost. Every piece of data of an ast node is optional, this is due to the fact that the parser is completely error tolerant.

Each representation has its advantages:

SyntaxNodes

Very simple traversing of the syntax tree through functions on them.

Easily able to convert to underlying text, range, or tokens.

Contain all whitespace bound to the underlying production (in the case of lossless parsing).

Can be easily converted into its typed representation with zero cost.

Can be turned into a pretty representation with fmt debug.

AST Nodes

Easy access to properties of the underlying production.

Zero cost conversion to a syntax node.

In conclusion, the use of both representations means we are not constrained to acting through typed nodes. Which makes traversal hard and you often have to resort to autogenerated visitor patterns. AST nodes are simply a way to easily access subproperties of a syntax node.event;

Modules

syntax

The Js syntax itself and parser functions.

Macros

token_set

Utility macro for making a new token set

Structs

LosslessTreeSink

Structure for converting events to a syntax tree representation, while preserving whitespace.

Parse

A utility struct for managing the result of a parser job

TokenSet

Enums

Event

Events emitted by the Parser, these events are later made into a syntax tree with process into TreeSink.

JsSyntaxFeature

LexContext

Context in which the lexer should lex the next token

ReLexContext

Context in which the LexContext’s current should be re-lexed.

Traits

TreeSink

An abstraction for syntax tree implementations

Functions

parse

Parses the provided string as a EcmaScript program using the provided syntax features.

parse_common

parse_expression

Losslessly Parse text into an expression Parse which can then be turned into an untyped root JsSyntaxNode. Or turned into a typed JsExpressionSnipped with tree.

parse_module

Same as parse_text but configures the parser to parse an ECMAScript module instead of a script

parse_script

Parse text into a Parse which can then be turned into an untyped root JsSyntaxNode. Or turned into a typed JsScript with tree.

process

Generate the syntax tree with the control of events.

Type Definitions

ParseDiagnostic

The type of error emitted by the parser, this includes warnings, notes, and errors. It also includes labels and possibly notes