Struct rome_js_parser::ParseDiagnostic

pub struct ParseDiagnostic { /* private fields */ }

A specialized diagnostic for the parser

Parser diagnostics are always errors.

A parser diagnostics structured in this way:

a mandatory message and a mandatory [TextRange]

a list of details, useful to give more information and context around the error

a hint, which should tell the user how they could fix their issue

These information are printed in this exact order.

Implementations

impl ParseDiagnostic

pub fn new(file_id: FileId, message: impl Display, span: impl AsSpan) -> Self

pub const fn is_error(&self) -> bool

pub fn detail(self, range: impl AsSpan, message: impl Display) -> Self

Use this API if you want to highlight more code frame, to help to explain where’s the error.

A detail is printed after the actual error and before the hint.

Examples

use rome_console::fmt::{Termcolor};
use rome_console::markup;
use rome_diagnostics::v2::{DiagnosticExt, FileId, PrintDiagnostic, console::fmt::Formatter};
use rome_js_parser::ParseDiagnostic;
use rome_js_syntax::TextRange;
use rome_rowan::TextSize;
use std::fmt::Write;

let source = "const a";
let range = TextRange::new(TextSize::from(0), TextSize::from(5));
let mut diagnostic = ParseDiagnostic::new(FileId::zero(), "this is wrong!", range)
    .detail(TextRange::new(TextSize::from(6), TextSize::from(7)), "This is reason why it's broken");

let mut write = rome_diagnostics::termcolor::Buffer::no_color();
let error = diagnostic
    .clone()
    .with_file_path(FileId::zero())
    .with_file_source_code(source.to_string());
Formatter::new(&mut Termcolor(&mut write))
    .write_markup(markup! {
    {PrintDiagnostic(&error)}
})
    .expect("failed to emit diagnostic");

let mut result = String::new();
write!(
    result,
    "{}",
    std::str::from_utf8(write.as_slice()).expect("non utf8 in error buffer")
).expect("");

let expected = r#"parse ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

  × this is wrong!
  
  > 1 │ const a
      │ ^^^^^
  
  i This is reason why it's broken
  
  > 1 │ const a
      │       ^
  
"#;
assert_eq!(result, expected);

pub fn hint(self, message: impl Display) -> Self

Small message that should suggest the user how they could fix the error

Hints are rendered a last part of the diagnostics

Examples

use rome_console::fmt::{Termcolor};
use rome_console::markup;
use rome_diagnostics::v2::{DiagnosticExt, FileId, PrintDiagnostic, console::fmt::Formatter};
use rome_js_parser::ParseDiagnostic;
use rome_js_syntax::TextRange;
use rome_rowan::TextSize;
use std::fmt::Write;

let source = "const a";
let range = TextRange::new(TextSize::from(0), TextSize::from(5));
let mut diagnostic = ParseDiagnostic::new(FileId::zero(), "this is wrong!", range)
    .hint("You should delete the code");

let mut write = rome_diagnostics::termcolor::Buffer::no_color();
let error = diagnostic
    .clone()
    .with_file_path(FileId::zero())
    .with_file_source_code(source.to_string());
Formatter::new(&mut Termcolor(&mut write))
    .write_markup(markup! {
    {PrintDiagnostic(&error)}
})
    .expect("failed to emit diagnostic");

let mut result = String::new();
write!(
    result,
    "{}",
    std::str::from_utf8(write.as_slice()).expect("non utf8 in error buffer")
).expect("");

let expected = r#"parse ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

  × this is wrong!
  
  > 1 │ const a
      │ ^^^^^
  
  i You should delete the code
  
"#;
assert_eq!(result, expected);